\name{Ranks}
\alias{Ranks}
\title{Ranks computation}
\description{Computes the ranks of the distances of individuals to grid points.}
\usage{Ranks(data)}
\arguments{
\item{data}{the list of the six items provided by the Data and the Mirror functions.}
}
\value{a vector of all the ranks of the distances of individuals to grid points.}
\seealso{\code{\link{Data}} for the format of the data variable.}
\keyword{manip}
