\name{wombsoft-internal}
\alias{Border}
\alias{BottomSide}
\alias{Conversion}
\alias{ConvexHull}
\alias{CopyCodominant}
\alias{CopyDominant}
\alias{DataEncoding}
\alias{DataEncodingCodominant}
\alias{InConvexHull}
\alias{InfPolar}
\alias{ListValues}
\alias{PlotDir}
\alias{PolarSort}
\alias{Quicksort}
\alias{Symmetry}
\alias{TopSide}
\alias{TurnLeft}
\title{Internal functions}
\description{
Internal functions
}
\usage{
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
